# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{id}/suppliers:
    get:
      operationId: getSupplierRecommendations
      tags: [suppliers]
      summary: Get green supplier recommendations for a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of supplier recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierRecommendation"
  /businesses/{id}/progress:
    get:
      operationId: getProgressHistory
      tags: [progress]
      summary: Get monthly progress records for a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Monthly progress history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressRecord"
    post:
      operationId: logProgress
      tags: [progress]
      summary: Log a monthly progress update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogProgressBody"
      responses:
        "201":
          description: Logged progress record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressRecord"
  /businesses/{id}/esg-report:
    get:
      operationId: getEsgReport
      tags: [businesses]
      summary: Get AI-generated ESG report for a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ESG report content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EsgReport"
        "404":
          description: No report yet
    post:
      operationId: generateEsgReport
      tags: [businesses]
      summary: Generate a new AI ESG report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Generated ESG report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EsgReport"
...
components:
  schemas:
...
    SupplierRecommendation:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        name:
          type: string
        category:
          type: string
          description: "energi | bahan_baku | packaging | logistik"
        description:
          type: string
        greenCertification:
          type: string
          nullable: true
        location:
          type: string
        estimatedEmissionReduction:
          type: number
          description: "Estimated CO2e reduction if switched to this supplier"
        reasoning:
          type: string
        website:
          type: string
          nullable: true
      required:
        - id
        - businessId
        - name
        - category
        - description
        - location
        - estimatedEmissionReduction
        - reasoning
    ProgressRecord:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        month:
          type: string
          description: "e.g. 2024-01"
        actualEmissions:
          type: number
          description: "Actual CO2e this month in tonnes"
        baselineEmissions:
          type: number
          description: "Baseline CO2e for comparison"
        reductionPercent:
          type: number
          description: "Percentage reduction vs baseline"
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - month
        - actualEmissions
        - baselineEmissions
        - reductionPercent
        - createdAt
    LogProgressBody:
      type: object
      properties:
        month:
          type: string
        actualEmissions:
          type: number
        baselineEmissions:
          type: number
        notes:
          type: string
      required:
        - month
        - actualEmissions
        - baselineEmissions
    EsgReport:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        period:
          type: string
        totalEmissions:
          type: number
        reductionFromBaseline:
          type: number
        environmentalScore:
          type: integer
        socialScore:
          type: integer
        governanceScore:
          type: integer
        overallScore:
          type: integer
        executiveSummary:
          type: string
        environmentalSection:
          type: string
        socialSection:
          type: string
        governanceSection:
          type: string
        recommendations:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - period
        - totalEmissions
        - reductionFromBaseline
        - environmentalScore
        - socialScore
        - governanceScore
        - overallScore
        - executiveSummary
        - environmentalSection
        - socialSection
        - governanceSection
        - recommendations
        - createdAt
...