# lib/api-zod/src/generated/types/esgReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EsgReport {
  id: number;
  businessId: number;
  period: string;
  totalEmissions: number;
  reductionFromBaseline: number;
  environmentalScore: number;
  socialScore: number;
  governanceScore: number;
  overallScore: number;
  executiveSummary: string;
  environmentalSection: string;
  socialSection: string;
  governanceSection: string;
  recommendations: string;
  createdAt: Date;
}