# lib/api-zod/src/generated/types/supplierRecommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SupplierRecommendation {
  id: number;
  businessId: number;
  name: string;
  /** energi | bahan_baku | packaging | logistik */
  category: string;
  description: string;
  greenCertification?: string | null;
  location: string;
  /** Estimated CO2e reduction if switched to this supplier */
  estimatedEmissionReduction: number;
  reasoning: string;
  website?: string | null;
}